# wafaa/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wafaa API Documentation",
        "description": "API documentation for Wafaa job portal application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/jobs": {
            "get": {
                "tags": [
                    "Offres d'emploi"
                ],
                "summary": "Lister toutes les offres d'emploi",
                "description": "Retourne la liste complète des offres d'emploi disponibles.",
                "operationId": "9538d30d530d4c3680ea31fe228944e2",
                "responses": {
                    "200": {
                        "description": "Liste des offres récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Job"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Offres d'emploi"
                ],
                "summary": "Créer une nouvelle offre d'emploi",
                "description": "Permet à un employeur connecté de créer une nouvelle offre d'emploi.",
                "operationId": "36deb2a6480382cf8385911c10aeebf0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "company",
                                    "location"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Développeur web"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Développement d'applications Laravel."
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "TechCorp"
                                    },
                                    "salary": {
                                        "type": "number",
                                        "example": 4000
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Casablanca"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Offre créée avec succès"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{id}": {
            "put": {
                "tags": [
                    "Offres d'emploi"
                ],
                "summary": "Mettre à jour une offre d'emploi",
                "operationId": "66762b778fbd20dd4fbd09628dabba17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Développeur backend"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Mise à jour des API REST."
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "TechCorp"
                                    },
                                    "salary": {
                                        "type": "number",
                                        "example": 4500
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Rabat"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Offre mise à jour avec succès"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Offre non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Offres d'emploi"
                ],
                "summary": "Supprimer une offre d'emploi",
                "operationId": "5537fd998abc3a62eef8c09e4ae52000",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Offre supprimée avec succès"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Offre non trouvée"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/search": {
            "get": {
                "tags": [
                    "Offres d'emploi"
                ],
                "summary": "Rechercher des offres d'emploi",
                "operationId": "0c0290a245c580348856b16cf6be0b81",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Filtrer par titre",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "company",
                        "in": "query",
                        "description": "Filtrer par entreprise",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filtrer par localisation",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats de la recherche"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Job": {
                "title": "Job",
                "description": "Modèle représentant une offre d'emploi",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Développeur Web"
                    },
                    "description": {
                        "type": "string",
                        "example": "Développer des applications Laravel."
                    },
                    "company": {
                        "type": "string",
                        "example": "TechCorp"
                    },
                    "salary": {
                        "type": "number",
                        "example": 4500
                    },
                    "location": {
                        "type": "string",
                        "example": "Casablanca"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-22T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-10-22T10:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utilisez un jeton Bearer pour vous authentifier",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}